🧪 [Test] Admin 공유일정 삭제 API 테스트 코드 작성

+++ gg-calendar-api/src/main/java/gg/calendar/api/admin/schedule/publicschedule/service/PublicScheduleAdminService.java
@@ -18,6 +18,7 @@ import gg.calendar.api.admin.schedule.publicschedule.controller.response.PublicS
 import gg.calendar.api.admin.schedule.publicschedule.controller.response.PublicScheduleAdminUpdateResDto;
 import gg.data.calendar.PublicSchedule;
 import gg.data.calendar.type.DetailClassification;
+import gg.data.calendar.type.ScheduleStatus;
 import gg.utils.dto.PageResponseDto;
 import gg.utils.exception.ErrorCode;
 import gg.utils.exception.custom.InvalidParameterException;
@@ -79,6 +80,9 @@ public class PublicScheduleAdminService {
 	public void deletePublicSchedule(Long id) {
 		PublicSchedule publicSchedule = publicScheduleAdminRepository.findById(id)
 			.orElseThrow(() -> new NotExistException(ErrorCode.PUBLIC_SCHEDULE_NOT_FOUND));
+		if (publicSchedule.getStatus().equals(ScheduleStatus.DELETE)) {
+			throw new InvalidParameterException(ErrorCode.PUBLIC_SCHEDULE_ALREADY_DELETED);
+		}
 		publicSchedule.delete();
 	}
